# [2/N] Chunk encoding optimization: bench mark the gnark based encoding