## Why are these changes needed?
To understand the performance of new chunk serialization and deserialization compared to existing approach.

Encoding (BenchmarkFrameGnarkSerialization) is 7x faster, decoding (BenchmarkFrameGnarkDeserialization) is 3.6x faster.

```
goos: linux
goarch: amd64
pkg: github.com/Layr-Labs/eigenda/encoding
cpu: Intel(R) Xeon(R) Platinum 8259CL CPU @ 2.50GHz
BenchmarkFrameGobSerialization-8       	   71583	     16657 ns/op
BenchmarkFrameGobSerialization-8       	   71329	     16743 ns/op
BenchmarkFrameGobSerialization-8       	   70735	     16912 ns/op
BenchmarkFrameGnarkSerialization-8     	  478899	      2402 ns/op
BenchmarkFrameGnarkSerialization-8     	  467806	      2390 ns/op
BenchmarkFrameGnarkSerialization-8     	  467905	      2402 ns/op
BenchmarkFrameGobDeserialization-8     	   32188	     34923 ns/op
BenchmarkFrameGobDeserialization-8     	   34275	     34864 ns/op
BenchmarkFrameGobDeserialization-8     	   33954	     34775 ns/op
BenchmarkFrameGnarkDeserialization-8   	  121190	      9715 ns/op
BenchmarkFrameGnarkDeserialization-8   	  122356	      9832 ns/op
BenchmarkFrameGnarkDeserialization-8   	  120322	      9718 ns/op
```


<!-- Please give a short summary of the change and the problem this solves. -->

## Checks

- [x] I've made sure the lint is passing in this PR.
- [x] I've made sure the tests are passing. Note that there might be a few flaky tests, in that case, please comment that they are not relevant.
- Testing Strategy
   - [ ] Unit tests
   - [ ] Integration tests
   - [ ] This PR is not tested :(